test_orphaned_upload_file_deletion
